Method for checking that line is in searched time range
